src/requirements.txt
llama-index
fastapi
python-dotenv
pydantic
openai
uvicorn